% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterStab.R
\name{clust.stab.index}
\alias{clust.stab.index}
\title{Perform clustering stability}
\usage{
clust.stab.index(data, nb.grs = 1:10, data.generator, clustering.method,
  clust.comp.score = NID, normalization = NID.normalizaton, nsim = 50,
  mc.cores = 3, options_clustering = list(),
  options_perturbation = list())
}
\arguments{
\item{data}{a data frame}
}
\description{
Perform clustering stability
}
