% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls_to_clustering.R
\name{clusterStab_gmm}
\alias{clusterStab_gmm}
\title{call to mclust formatted for the clusterStability package}
\usage{
clusterStab_gmm(data, vec_ngroup, ...)
}
\description{
call to mclust formatted for the clusterStability package
}
