% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls_to_clustering.R
\name{clusterStab_hclust}
\alias{clusterStab_hclust}
\title{call to hclust formatted for the clusterStability package}
\usage{
clusterStab_hclust(data, vec_ngroup, ...)
}
\description{
call to hclust formatted for the clusterStability package
}
