% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls_to_clustering.R
\name{clusterStab_kmeans}
\alias{clusterStab_kmeans}
\title{call to kmeans formatted for the clusterStability package}
\usage{
clusterStab_kmeans(data, vec_ngroup, options_clustering = list())
}
\description{
call to kmeans formatted for the clusterStability package
}
