% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustRstab-package.R
\docType{package}
\name{clustRstab-package}
\alias{clustRstab}
\alias{clustRstab-package}
\title{clustRstab: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Martina Sundqvist \email{martina.sundqvist@agroparistech.fr}

Authors:
\itemize{
  \item Julien Chiquet \email{julien.chiquet@inrae.fr} (\href{https://orcid.org/0000-0002-3629-3429}{ORCID})
  \item Guillem Rigaill \email{guillem.rigaill@inrae.fr}
}

}
\keyword{internal}
