% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustRstab.R
\name{clustRstab}
\alias{clustRstab}
\title{A function to compute cluster stability}
\usage{
clustRstab(
  data,
  kVec,
  typeOfComp = "all",
  perturbedDataFun = clustRstab::subsample,
  clAlgo = clustRstab::clAlgoKmeans,
  clCompScore = aricode::MARI,
  nsim = 100,
  baseLineCorrection = FALSE,
  plot = TRUE,
  nProp = 0.7,
  pProp = 0.9,
  noiseGaussianMean = 0,
  noiseGaussianSD = 1,
  randProjMethod = "Haar",
  randProjDim = 10
)
}
\arguments{
\item{data}{a data frame dsfdsf dqsdqs}

\item{kVec}{a vector for the successive number of group considered}

\item{perturbedDataFun}{a function for perturbing the original data}

\item{clAlgo}{a function: the clustering method used}

\item{clCompScore}{dqsds}

\item{nsim}{a scalar, the number of simulation performed to estimate the clustergin stability}

\item{baseLineCorrection}{dada}

\item{plot}{dazadz}

\item{nProp}{azdzd}

\item{pProp}{fsdfsd}
}
\description{
clustRstab() computes cluster stability as the arithmetic mean of the cluster comparision
scores. It also computes sd, min, et max du score (by K). It makes a simple plot
}
