% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterStab.R
\name{instab.index2}
\alias{instab.index2}
\title{Perform clustering stability}
\usage{
instab.index2(data, nb.grs = 1:10, props = c(1, 1), clustering.method,
  nsim = 100, mc.cores = 3, score = "NID", ...)
}
\arguments{
\item{data}{a data frame}
}
\description{
Perform clustering stability
}
