% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCI60.R
\docType{data}
\name{NCI60}
\alias{NCI60}
\title{NCI60}
\format{
A list with two entries:
\describe{
  \item{expr}{a data.frame with 6830 rows and 64 columns of expression levels}
  \item{type}{a character vector containing the cancer types.}
  ...
}
}
\source{
\url{http://genome-www.stanford.edu/nci60/}
}
\usage{
NCI60
}
\description{
Microarray Gene Expression Profiles Of The NCI 60 Cell Lines
}
\details{
The data contains expression levels on 6830 genes from 64 cancer cell lines. Cancer type is also recorded.
}
\keyword{datasets}
