% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.R
\name{score.pairs}
\alias{score.pairs}
\title{Generate perturbed data sets by subsampling varibles}
\usage{
score.pairs(cl1, cl2, mc.cores)
}
\description{
Generate perturbed data sets by subsampling varibles
}
