% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbedDataFuns.R
\name{subSample}
\alias{subSample}
\title{A function to subsample a dataset}
\usage{
subSample(data, nProp, pProp, ...)
}
\arguments{
\item{data}{A n x p data.frame() with n the number of observations and p the number of varaibles}

\item{nProp}{The proportion of subsampled observations}

\item{pProp}{The proportion of subsampled variables}
}
\description{
subSample() subsamples nProp observation and pProp variables of a n x p data.frame
}
