% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perturbed_data.R
\name{subVar_data}
\alias{subVar_data}
\title{Generate perturbed data sets by subsampling varibles}
\usage{
subVar_data(i, data, param)
}
\description{
Generate perturbed data sets by subsampling varibles
}
