% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterStab.R
\name{clust.stab.index}
\alias{clust.stab.index}
\title{Perform clustering stability}
\usage{
clust.stab.index(data, nb.grs = 1:10, data.generator, data.genenerator.param,
  clustering.method, nsim = 100, mc.cores = 3, score = "NID",
  options_clustering = list(), options_perturbation = list())
}
\arguments{
\item{data}{a data frame}
}
\description{
Perform clustering stability
}
